simulation: support better linear constraints
Linear constraints worked well for triangles but for quads, the minimum energy is formed when the area of the quad becomes zero. To prevent the collapse of the quad, add constraints for the diagnols of the quad. This also means that only triangles and quads are support as of right now.