text: hack: hack scale multiplier for FUnits to px
Without the hack scale multiplier, the scale multiplier is 1.21 times the actual value it would be, tested by having a 72pt text rendered on primary monitor and checking with a scale to see if the box is 1 inch tall.

The factor by which it is off is very close to the face.height/units_per_em.
Thus as a hack, we multiply the base multiplier with units_per_em/face.height.

This may or may not be correct approach but it works.

Another point to note is that the system dpi reported using `xdpyinfo | grep -B 2 resolution` is 96 but when manually calculating the dpi value, it is 141.76.

For text rendered to be the same size as other system software text, the dpi from the system is taken without the hack multiplier. It is very surprising that this is the case since it doesn't lead to a 72pt font being 1 inch tall.